Use phpactor--command-argments in company-phpactor--get-suggestions
Previously, the function `company-phpactor--get-suggestions` took two
variables, SOURCE and OFFSET, but didn't even use them (I forgot to
implement the changes). Now the function uses
`phpactor--command-argments`

What worries me now is that I think this makes it slightly less
testable? But perhaps the package is small enough that it won't be an issue.