support for the "references replace" rpc action
add support for the "references" rpc action in "replace" mode and
support for "files_references" action within emacs.

This refactoring finds and replaces the references with the given
replacement string, and ends with "file_references" to inform the user
of the changed made by phpactor.

This commit also adds a "phpactor-list-references" function which
opens a buffer showing the list of changes initiated from the buffer
this function was invoked.