[IMP] Make `phpactor-import-class' actually interactive
Despite being declared as `(interactive)`, the function can't be called
as such since there's no NAME provided to it. This patch aims to improve
the usability of this function, since now explicit `NAME` arg could be
omitted.

As of now, the function looks up for NAME in this particular order:

* a NAME argument explicitly passed to function -- to use in code,
* contents of current region, if any -- interactive usage,
* a symbol on a current cursor position -- interactive usage.